Updated everything to use Semantic Kernel's new ChatGPT support. Removed Azure OpenAI chat support for now (will add back in soon after SK supports it).
