--- src/ConversationalSpeaker/Handlers/OpenAiModels/OpenAiChatMessage.cs
@@ -1,8 +1,0 @@
-﻿namespace ConversationalSpeaker.Handlers.OpenAiModels
-{
-    public class OpenAIChatMessage
-    {
-        public string role { get; set; }
-        public string content { get; set; }
-    }
-}
