+++ src/ConversationalSpeaker/configuration.json
@@ -22,25 +22,13 @@
     "WakePhraseModel": "hey-computer.table"
   },
 
-  "AzureOpenAI": {
-    "Key": "", // dotnet user-secrets set "AzureOpenAI:Key" "****"
-    "Endpoint": "", // dotnet user-secrets set "AzureOpenAI:Endpoint" "****"
-    "Deployment": "gpt-35-turbo",
-    "MaxTokens": 4095,
-    "Temperature": 0.7, // 0.0 - 1.0
-    "TopP": 1, // 0.0 - 1.0
-    "FrequencyPenalty": 0, // 0.0 - 2.0
-    "PresencePenalty": 0, // 0.0 - 2.0
-    "Model": "gpt-35-turbo" // currently optimized for "chat" models only
-  },
-
   "OpenAI": {
     "Key": "", // dotnet user-secrets set "OpenAI:Key" "****"
     "MaxTokens": 1500,
     "Temperature": 0.7, // 0.0 - 1.0
     "FrequencyPenalty": 0, // 0.0 - 2.0
     "PresencePenalty": 0, // 0.0 - 2.0
-    "Model": "gpt-3.5-turbo", // For a list of available models, see https://beta.openai.com/docs/models/overview, currently optimized for "chat" models only
-    "OrganizationId": "" // Organization ID for those with multiple organizations (optional).
+    "Model": "gpt-3.5-turbo", // For a list of available models, see https://beta.openai.com/docs/models/overview
+    //"OrganizationId": "" // Organization ID for those with multiple organizations (optional).
   }
 }
